refactor: remove ProfileCard from ProfilePage to EditableProfileCard
