+++ public/locales/en/translation.json
@@ -64,5 +64,7 @@
     "Recommendations": "Recommendations",
     "Create a new article": "Create a new article",
     "Editing an article with ID": "Editing an article with ID ",
-    "Error loading articles": "Error loading articles"
+    "Error loading articles": "Error loading articles",
+    "Profile not found": "Profile not found"
+
 }
