+++ public/locales/ru/translation.json
@@ -72,6 +72,6 @@
 "Recommendations": "Рекомендуем",
 "Create a new article": "Создание новой статьи",
 "Editing an article with ID": "Редактирование статьи с ID ",
-"Error loading articles": "Ошибка при загрузке статей"
-
+"Error loading articles": "Ошибка при загрузке статей",
+"Profile not found": "Профиль не найден"
 }
