+++ src/app/providers/StoreProvider/config/StateSchema.ts
@@ -4,7 +4,6 @@ import {
 import { AxiosInstance } from 'axios';
 import { AddCommentFormSchema } from '../../../../features/addCommentForm';
 import { ArticleDetailsSchema } from '../../../../entities/Article';
-import { ProfileSchema } from '../../../../entities/Profile';
 import { LoginSchema } from '../../../../features/AuthByUsername/ui';
 import { CounterSchema } from '../../../../entities/Counter';
 import { UserSchema } from '../../../../entities/User';
@@ -12,6 +11,7 @@ import { ArticlesPageSchema } from '../../../../pages/ArticlesPage';
 import { ScrollRestorationSchema } from '../../../../features/ScrollRestoration';
 import { ArticleDetailsPageSchema } from '../../../../pages/ArticleDetailsPage';
 import { rtkApi } from '../../../../shared/api/rtkApi';
+import { ProfileSchema } from '../../../../features/editableProfileCard';
 
 export interface StateSchema {
     counter: CounterSchema;
