+++ src/entities/Profile/index.ts
@@ -1,22 +1,3 @@
-export {
-    Profile,
-    ProfileSchema,
-    ValidateProfileError,
-} from './model/types/profile';
-
-export {
-    profileActions,
-    profileReducer,
-} from './model/slice/profileSlice';
-
-export { fetchProfileData } from './model/services/fetchProfileData/fetchProfileData';
-export { updateProfileData } from './model/services/updateProfileData/updateProfileData';
+export { Profile } from './model/types/profile';
 
 export { ProfileCard } from './ui/ProfileCard/ProfileCard';
-
-export { getProfileData } from '../../entities/Profile/model/selectors/getProfileData/getProfileData';
-export { getProfileIsLoading } from '../../entities/Profile/model/selectors/getProfileIsLoading/getProfileIsLoading';
-export { getProfileError } from '../../entities/Profile/model/selectors/getProfileError/getProfileError';
-export { getProfileReadonly } from '../../entities/Profile/model/selectors/getProfileReadonly/getProfileReadonly';
-export { getProfileForm } from '../../entities/Profile/model/selectors/getProfileForm/getProfileForm';
-export { getProfileValidateErrors } from '../../entities/Profile/model/selectors/getProfileValidateErrors/getProfileValidateErrors';
