+++ src/entities/Profile/model/types/profile.ts
@@ -1,13 +1,6 @@
 import { Country } from '../../../../entities/Country';
 import { Currency } from '../../../../entities/Currency';
 
-export enum ValidateProfileError {
-    INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
-    INCORRECT_AGE = 'INCORRECT_AGE',
-    INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
-    NO_DATA = 'NO_DATA',
-    SERVER_ERROR = 'SERVER_ERROR',
-}
 export interface Profile {
     id?: string;
     first?: string;
@@ -18,13 +11,4 @@ export interface Profile {
     city?: string,
     username?: string;
     avatar?: string;
-}
-
-export interface ProfileSchema {
-    data?: Profile;
-    form?: Profile;
-    isLoading: boolean;
-    error?: string;
-    readonly: boolean;
-    validateErrors?: ValidateProfileError[];
 }
