+++ src/features/editableProfileCard/index.ts
@@ -1,0 +1,2 @@
+export { EditableProfileCard } from './ui/EditableProfileCard/EditableProfileCard';
+export { ProfileSchema } from './model/types/editableProfileCardSchema';
