+++ src/features/editableProfileCard/model/selectors/getProfileData/getProfileData.test.ts
