+++ src/features/editableProfileCard/model/selectors/getProfileData/getProfileData.ts
@@ -1,3 +1,3 @@
-import { StateSchema } from 'app/providers/StoreProvider';
+import { StateSchema } from '../../../../../app/providers/StoreProvider';
 
 export const getProfileData = (state: StateSchema) => state.profile?.data;
