+++ src/features/editableProfileCard/model/selectors/getProfileError/getProfileError.ts
@@ -1,3 +1,3 @@
-import { StateSchema } from 'app/providers/StoreProvider';
+import { StateSchema } from '../../../../../app/providers/StoreProvider';
 
 export const getProfileError = (state: StateSchema) => state.profile?.error;
