+++ src/features/editableProfileCard/model/selectors/getProfileIsLoading/getProfileIsLoading.ts
@@ -1,3 +1,3 @@
-import { StateSchema } from 'app/providers/StoreProvider';
+import { StateSchema } from '../../../../../app/providers/StoreProvider';
 
 export const getProfileIsLoading = (state: StateSchema) => state.profile?.isLoading;
