+++ src/features/editableProfileCard/model/selectors/getProfileValidateErrors/getProfileValidateErrors.test.ts
@@ -1,5 +1,5 @@
 import { StateSchema } from '../../../../../app/providers/StoreProvider';
-import { ValidateProfileError } from '../../types/profile';
+import { ValidateProfileError } from '../../types/editableProfileCardSchema';
 import { getProfileValidateErrors } from './getProfileValidateErrors';
 
 describe('getProfileValidateErrors.test', () => {
