+++ src/features/editableProfileCard/model/selectors/getProfileValidateErrors/getProfileValidateErrors.ts
@@ -1,3 +1,3 @@
-import { StateSchema } from 'app/providers/StoreProvider';
+import { StateSchema } from '../../../../../app/providers/StoreProvider';
 
 export const getProfileValidateErrors = (state: StateSchema) => state.profile?.validateErrors;
