+++ src/features/editableProfileCard/model/services/fetchProfileData/fetchProfileData.test.ts
