+++ src/features/editableProfileCard/model/services/fetchProfileData/fetchProfileData.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
+import { Profile } from '../../../../../entities/Profile';
 import { ThunkConfig } from '../../../../../app/providers/StoreProvider';
-import { Profile } from '../../types/profile';
 
 export const fetchProfileData = createAsyncThunk<
     Profile,
