+++ src/features/editableProfileCard/model/services/updateProfileData/updateProfileData.ts
@@ -1,7 +1,8 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
+import { Profile } from '../../../../../entities/Profile';
 import { ThunkConfig } from '../../../../../app/providers/StoreProvider';
-import { Profile, ValidateProfileError } from '../../types/profile';
 import { getProfileForm } from '../../selectors/getProfileForm/getProfileForm';
+import { ValidateProfileError } from '../../types/editableProfileCardSchema';
 import { validateProfileData } from '../validateProfileData/validateProfileData';
 
 export const updateProfileData = createAsyncThunk<
