+++ src/features/editableProfileCard/model/slice/profileSlice.test.ts
@@ -1,8 +1,8 @@
 import { profileActions, profileReducer } from './profileSlice';
 import { Currency } from '../../../../entities/Currency';
 import { Country } from '../../../../entities/Country';
-import { ProfileSchema, ValidateProfileError } from '../types/profile';
 import { updateProfileData } from '../services/updateProfileData/updateProfileData';
+import { ProfileSchema, ValidateProfileError } from '../types/editableProfileCardSchema';
 
 const data = {
     username: 'admin',
