+++ src/features/editableProfileCard/ui/EditableProfileCardHeader/EditableProfileCardHeader.tsx
@@ -1,21 +1,22 @@
 import { useTranslation } from 'react-i18next';
+import { memo, useCallback } from 'react';
 import { useSelector } from 'react-redux';
-import { useCallback } from 'react';
-import { HStack } from '../../../../shared/ui/Stack';
-import { getUserAuthData } from '../../../../entities/User';
-import {
-    getProfileData, getProfileReadonly, profileActions, updateProfileData,
-} from '../../../../entities/Profile';
-import { Button, ButtonTheme } from '../../../../shared/ui/Button/Button';
-import { Text } from '../../../../shared/ui/Text/Text';
 import { classNames } from '../../../../shared/lib/classNames/classNames';
 import { useAppDispatch } from '../../../../shared/lib/hooks/useAppDispatch/useAppDispatch';
+import { HStack } from '../../../../shared/ui/Stack';
+import { Text } from '../../../../shared/ui/Text/Text';
+import { Button, ButtonTheme } from '../../../../shared/ui/Button/Button';
+import { getUserAuthData } from '../../../../entities/User';
+import { profileActions } from '../../model/slice/profileSlice';
+import { getProfileReadonly } from '../../model/selectors/getProfileReadonly/getProfileReadonly';
+import { getProfileData } from '../../model/selectors/getProfileData/getProfileData';
+import { updateProfileData } from '../../model/services/updateProfileData/updateProfileData';
 
-interface ProfilePageHeaderProps {
+interface EditableProfileCardHeaderProps {
     className?: string;
 }
 
-export const ProfilePageHeader = (props: ProfilePageHeaderProps) => {
+export const EditableProfileCardHeader = memo((props: EditableProfileCardHeaderProps) => {
     const {
         className,
     } = props;
@@ -74,4 +75,4 @@ export const ProfilePageHeader = (props: ProfilePageHeaderProps) => {
 
         </HStack>
     );
-};
+});
