+++ src/pages/ArticleDetailsPage/ui/ArticleDetailsPageHeader/ArticleDetailsPageHeader.stories.tsx
@@ -1,23 +1,17 @@
-// import { ComponentStory, ComponentMeta } from '@storybook/react';
-// import { Theme } from '../../../../app/providers/themeProvider';
-// import { ThemeDecorator } from '../../../../shared/config/storybook/ThemeDecorator';
-// import { ArticleDetailsPageHeader } from './ArticleDetailsPageHeader';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { ArticleDetailsPageHeader } from './ArticleDetailsPageHeader';
 
-// export default {
-//     title: 'pages/ArticleDetailsPageHeader',
-//     component: ArticleDetailsPageHeader,
-//     argTypes: {
-//         backgroundColor: { control: 'color' },
-//     },
-// } as ComponentMeta<typeof ArticleDetailsPageHeader>;
+export default {
+    title: 'pages/ArticleDetailsPage/ArticleDetailsPageHeader',
+    component: ArticleDetailsPageHeader,
+    argTypes: {
+        backgroundColor: { control: 'color' },
+    },
+} as ComponentMeta<typeof ArticleDetailsPageHeader>;
 
-// const Template: ComponentStory<typeof ArticleDetailsPageHeader> = (args) => <ArticleDetailsPageHeader {...args} />;
+const Template: ComponentStory<typeof ArticleDetailsPageHeader> = (args) => <ArticleDetailsPageHeader {...args} />;
 
-// export const Normal = Template.bind({});
-// Normal.args = {};
-
-// export const Dark = Template.bind({});
-// Dark.args = {};
-// Dark.decorators = [ThemeDecorator(Theme.DARK)];
+export const Normal = Template.bind({});
+Normal.args = {};
 
 // TO DO: FIX STORIES
