+++ src/pages/ArticleEditPage/ui/ArticleEditPage/ArticleEditPage.stories.tsx
@@ -1,48 +1,15 @@
-// import { ComponentStory, ComponentMeta } from '@storybook/react';
-// import { ThemeDecorator } from '../../../../shared/config/storybook/ThemeDecorator';
-// import { Theme } from '../../../../app/providers/themeProvider';
-// import ArticleEditPage from './ArticleEditPage';
-// import { StoreDecorator } from 'shared/config/storybook/StoreDecorator';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import ArticleEditPage from './ArticleEditPage';
 
-// export default {
-//     title: 'pages/ArticleEditPage',
-//     component: ArticleEditPage,
-//     argTypes: {
-//         backgroundColor: { control: 'color' },
-//     },
-// } as ComponentMeta<typeof ArticleEditPage>;
+export default {
+    title: 'pages/ArticleEditPage/ArticleEditPage',
+    component: ArticleEditPage,
+    argTypes: {
+        backgroundColor: { control: 'color' },
+    },
+} as ComponentMeta<typeof ArticleEditPage>;
 
-// const Template: ComponentStory<typeof ArticleEditPage> = (args) => <ArticleEditPage {...args} />;
+const Template: ComponentStory<typeof ArticleEditPage> = (args) => <ArticleEditPage {...args} />;
 
-// export const Normal = Template.bind({});
-// Normal.args = {};
-// Normal.decorators = [StoreDecorator({
-//     // articleDetailsPage: {
-
-//     // }
-//     // articleDetails: {
-//     //     data: article,
-//     // },
-// })];
-
-// export const Dark = Template.bind({});
-// Dark.args = {};
-// Dark.decorators = [ThemeDecorator(Theme.DARK)];
-
-// export const Loading = Template.bind({});
-// Loading.args = {};
-// Loading.decorators = [StoreDecorator({
-//     articleDetails: {
-//         isLoading: true,
-//     },
-// })];
-
-// export const Error = Template.bind({});
-// Error.args = {};
-// Error.decorators = [StoreDecorator({
-//     articleDetails: {
-//         error: 'error',
-//     },
-// })];
-
-// TO DO: FIX STORIES
+export const Normal = Template.bind({});
+Normal.args = {};
