+++ src/pages/ArticlesPage/ui/ArticlesPageFilters/ArticlesPageFilters.stories.tsx
@@ -1,17 +1,15 @@
-// import { ComponentStory, ComponentMeta } from '@storybook/react';
-// import { ArticlesPageFilters } from './ArticlesPageFilters';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { ArticlesPageFilters } from './ArticlesPageFilters';
 
-// export default {
-//     title: 'shared/ArticlesPageFilters',
-//     component: ArticlesPageFilters,
-//     argTypes: {
-//         backgroundColor: { control: 'color' },
-//     },
-// } as ComponentMeta<typeof ArticlesPageFilters>;
+export default {
+    title: 'pages/ArticlesPage/ArticlesPageFilters',
+    component: ArticlesPageFilters,
+    argTypes: {
+        backgroundColor: { control: 'color' },
+    },
+} as ComponentMeta<typeof ArticlesPageFilters>;
 
-// const Template: ComponentStory<typeof ArticlesPageFilters> = (args) => <ArticlesPageFilters {...args} />;
+const Template: ComponentStory<typeof ArticlesPageFilters> = (args) => <ArticlesPageFilters {...args} />;
 
-// export const Normal = Template.bind({});
-// Normal.args = {
-
-// };
+export const Normal = Template.bind({});
+Normal.args = {};
