+++ src/pages/ProfilePage/ui/ProfilePage.tsx
@@ -1,111 +1,29 @@
-import { useTranslation } from 'react-i18next';
-import { useCallback } from 'react';
-import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+import { Text } from '../../../shared/ui/Text/Text';
+import { EditableProfileCard } from '../../../features/editableProfileCard';
 import { VStack } from '../../../shared/ui/Stack';
-import { useInitialEffect } from '../../../shared/lib/hooks/useInitialEffect/useInitialEffect';
-import { Text, TextTheme } from '../../../shared/ui/Text/Text';
-import { useAppDispatch } from '../../../shared/lib/hooks/useAppDispatch/useAppDispatch';
 import { classNames } from '../../../shared/lib/classNames/classNames';
-import { DynamicModuleLoader, ReducersList } from '../../../shared/lib/components/DynamicModuleLoader/DynamicModuleLoader';
-import {
-    fetchProfileData, getProfileError, getProfileForm, getProfileIsLoading, getProfileReadonly, getProfileValidateErrors,
-    profileActions, ProfileCard, profileReducer, ValidateProfileError,
-} from '../../../entities/Profile';
-import { ProfilePageHeader } from './ProfilePageHeader/ProfilePageHeader';
-import { Country } from '../../../entities/Country';
-import { Currency } from '../../../entities/Currency';
 import { Page } from '../../../widgets/Page/Page';
-
-const reducers: ReducersList = {
-    profile: profileReducer,
-};
 
 interface ProfilePageProps {
     className?: string;
 }
 
 const ProfilePage = ({ className }: ProfilePageProps) => {
+    const { id } = useParams<{ id: string }>();
     const { t } = useTranslation();
-    const dispatch = useAppDispatch();
-    const formData = useSelector(getProfileForm);
-    const isLoading = useSelector(getProfileIsLoading);
-    const error = useSelector(getProfileError);
-    const readonly = useSelector(getProfileReadonly);
-    const validateErrors = useSelector(getProfileValidateErrors);
-    const { id } = useParams<{id: string}>();
 
-    const validateErrorTranslates = {
-        [ValidateProfileError.SERVER_ERROR]: t('SERVER_ERROR'),
-        [ValidateProfileError.INCORRECT_COUNTRY]: t('INCORRECT_COUNTRY'),
-        [ValidateProfileError.NO_DATA]: t('NO_DATA'),
-        [ValidateProfileError.INCORRECT_USER_DATA]: t('INCORRECT_USER_DATA'),
-        [ValidateProfileError.INCORRECT_AGE]: t('INCORRECT_AGE'),
-    };
-
-    useInitialEffect(() => {
-        if (id) {
-            dispatch(fetchProfileData(id));
-        }
-    });
-
-    const onChangeFirstName = useCallback((value?: string) => {
-        dispatch(profileActions.updateProfile({ first: value || '' }));
-    }, [dispatch]);
-
-    const onChangeLastName = useCallback((value?: string) => {
-        dispatch(profileActions.updateProfile({ lastname: value || '' }));
-    }, [dispatch]);
-
-    const onChangeCity = useCallback((value?: string) => {
-        dispatch(profileActions.updateProfile({ city: value || '' }));
-    }, [dispatch]);
-
-    const onChangeAge = useCallback((value?: string) => {
-        dispatch(profileActions.updateProfile({ age: Number(value || 0) }));
-    }, [dispatch]);
-
-    const onChangeUsername = useCallback((value?: string) => {
-        dispatch(profileActions.updateProfile({ username: value || '' }));
-    }, [dispatch]);
-
-    const onChangeAvatar = useCallback((value?: string) => {
-        dispatch(profileActions.updateProfile({ avatar: value || '' }));
-    }, [dispatch]);
-
-    const onChangeCurrency = useCallback((currency: Currency) => {
-        dispatch(profileActions.updateProfile({ currency }));
-    }, [dispatch]);
-
-    const onChangeCountry = useCallback((country: Country) => {
-        dispatch(profileActions.updateProfile({ country }));
-    }, [dispatch]);
+    if (!id) {
+        return <Text text={t('Profile not found')} />;
+    }
 
     return (
-        <DynamicModuleLoader reducers={reducers} removeAfterUnmount>
-            <Page className={classNames('', {}, [className])}>
-                <VStack gap="16" max>
-                    <ProfilePageHeader />
-                    {validateErrors?.length && validateErrors.map((error) => (
-                        <Text key={error} theme={TextTheme.ERROR} text={validateErrorTranslates[error]} />
-                    ))}
-                    <ProfileCard
-                        data={formData}
-                        isLoading={isLoading}
-                        error={error}
-                        readonly={readonly}
-                        onChangeFirstName={onChangeFirstName}
-                        onChangeLastName={onChangeLastName}
-                        onChangeAge={onChangeAge}
-                        onChangeCity={onChangeCity}
-                        onChangeUsername={onChangeUsername}
-                        onChangeAvatar={onChangeAvatar}
-                        onChangeCurrency={onChangeCurrency}
-                        onChangeCountry={onChangeCountry}
-                    />
-                </VStack>
-            </Page>
-        </DynamicModuleLoader>
+        <Page className={classNames('', {}, [className])}>
+            <VStack gap="16" max>
+                <EditableProfileCard id={id} />
+            </VStack>
+        </Page>
     );
 };
