+++ src/shared/config/storybook/StoreDecorator.tsx
@@ -2,12 +2,12 @@ import { ReducersMapObject } from '@reduxjs/toolkit';
 import { Story } from '@storybook/react';
 import { ReducersList } from '../../../shared/lib/components/DynamicModuleLoader/DynamicModuleLoader';
 import { StateSchema, StoreProvider } from '../../../app/providers/StoreProvider';
-import { profileReducer } from '../../../entities/Profile';
 import { loginReducer } from '../../../features/AuthByUsername/model/slice/LoginSlice';
 import { articleDetailsReducer } from '../../../entities/Article/model/slice/articleDetailsSlice';
 import { addCommentFormReducer } from '../../../features/addCommentForm/model/slices/addCommentFormSlice';
 import { articleDetailsPageReducer } from '../../../pages/ArticleDetailsPage/model/slices';
 import { articlesPageReducer } from '../../../pages/ArticlesPage/model/slices/articlesPageSlice';
+import { profileReducer } from '../../../features/editableProfileCard/model/slice/profileSlice';
 
 const defaultAsyncReducers: ReducersList = {
     LoginForm: loginReducer,
